requirements.txt
gunicorn
flask

flask-wtf
itsdangerous
flask-mail
pymysql
sqlalchemy
flask_sqlalchemy
sqlalchemy-datatables

pandas
